+++ src/volatility_trading/backtesting/reporting/schemas.py
@@ -46,6 +46,7 @@ class BacktestReportBundle:
     summary_metrics: SummaryMetrics
     equity_and_drawdown: pd.DataFrame
     trades: pd.DataFrame
+    entry_stress_diagnostics: pd.DataFrame
     exposures_daily: pd.DataFrame
     margin_diagnostics_daily: pd.DataFrame
     rolling_metrics: pd.DataFrame